+++ CHANGES.md
@@ -9,3 +9,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 
 - Initial Project Setup with Lint, Test and Bundle Setup
+- Basic Editor Components Setup (Editor, EditorBlock and Example Components)